put fallback module in its own file
